requirements.txt
# Multi-Agent Systems Lab - Unified Requirements
# Combined requirements for both AutoGen and CrewAI frameworks
# Install with: pip install -r requirements.txt

# Core Frameworks
pyautogen>=0.2.0              # Microsoft AutoGen framework
crewai>=0.1.0                # CrewAI framework
crewai-tools>=0.0.1          # CrewAI tools extension

# API & LLM
openai>=1.0.0                # OpenAI API client
python-dotenv>=1.0.0         # Environment variable management

# Utilities
requests>=2.31.0             # HTTP library
pydantic>=2.0.0              # Data validation
